feat(deps): lazy load <DonationDialogContent>
